Prototype Finished
The hamburger menu now additionaly allows the user to select the employees sheet.

Hints are now determined by the currently selected sheet, variables for the 4 column names are defined in the methods within SheetRepository which select which select values for each sheet.
These methods are called within the hamburger menu's selection within SingleFragmentActivity.

Sheets who's final column is not a date do not currently work for the listview.